Move tests around.
