+++ distribution/lib/Standard/Table/0.0.0-dev/package.yaml
@@ -15,7 +15,6 @@ component-groups:
           - Standard.Table.Data.Table.Table.new
           - Standard.Table.Data.Table.Table.from_rows
           - Standard.Table.Data.Column.Column.from_vector
-          - Standard.Base.System.File.read
     - Standard.Base.Select:
         exports:
           - Standard.Table.Data.Table.Table.at
