+++ distribution/lib/Standard/Table/0.0.0-dev/src/Main.enso
@@ -13,8 +13,7 @@ import project.Data.Match_Columns.Match_Columns
 import project.Data.Position.Position
 import project.Data.Aggregate_Column.Aggregate_Column
 
-import project.IO.File_Read
-import project.IO.Auto_Detect.Auto_Detect
+import project.IO.Auto_Detect
 
 import project.Delimited.Quote_Style.Quote_Style
 import project.Delimited.Delimited_Format
@@ -34,8 +33,7 @@ export project.Data.Match_Columns.Match_Columns
 export project.Data.Position.Position
 export project.Data.Aggregate_Column.Aggregate_Column
 
-export project.IO.File_Read
-export project.IO.Auto_Detect.Auto_Detect
+export project.IO.Auto_Detect
 
 export project.Delimited.Quote_Style.Quote_Style
 from project.Delimited.Delimited_Format export Delimited_Format, Delimited
