+++ std-bits/base/src/main/java/org/enso/base/file_format/ByteFormatSPI.java
@@ -1,7 +1,7 @@
-package org.enso.base;
+package org.enso.base.file_format;
 
 @org.openide.util.lookup.ServiceProvider(service = FileFormatSPI.class)
-public class TextFormatSPI  extends FileFormatSPI {
+public class ByteFormatSPI  extends FileFormatSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Base.System.File_Format";
@@ -9,6 +9,6 @@ public class TextFormatSPI  extends FileFormatSPI {
 
   @Override
   protected String getTypeName() {
-    return "Plain_Text_Format";
+    return "Bytes";
   }
 }
