+++ std-bits/table/src/main/java/org/enso/table/read/ExcelFormatSPI.java
@@ -1,6 +1,6 @@
 package org.enso.table.read;
 
-import org.enso.base.FileFormatSPI;
+import org.enso.base.file_format.FileFormatSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = FileFormatSPI.class)
 public class ExcelFormatSPI extends FileFormatSPI {
