+++ test/Table_Tests/src/Main.enso
@@ -4,11 +4,9 @@ from Standard.Test import Test_Suite
 
 import project.In_Memory_Tests
 import project.Database.Main as Database_Tests
-import project.File_Read_Spec
 import project.Data_Formatter_Spec
 
 main = Test_Suite.run_main <|
     In_Memory_Tests.in_memory_spec
     Database_Tests.databases_spec
-    File_Read_Spec.spec
     Data_Formatter_Spec.spec
