+++ test/Tests/src/System/File_Read_Spec.enso
@@ -1,7 +1,6 @@
 from Standard.Base import all
 from Standard.Base.System.File_Format import Unsupported_File_Type_Data
 
-import Standard.Table.IO.File_Read
 from Standard.Test import Test, Test_Suite, Problems
 
 spec =
